Move L&P test cases and utility classes to new package
