+++ docs/performance-testing.md
@@ -2,7 +2,7 @@
 
 TEAMMATES makes use of [JMeter](https://jmeter.apache.org/) for load and performance (L&P) testing, and uses the [JMeter Java API](https://jmeter.apache.org/api/index.html) and TestNG to automate the process of running performance tests.
 
-The performance test cases are located in the [`teammates.e2e.cases.lnp`](https://github.com/TEAMMATES/teammates/tree/master/src/e2e/java/teammates/e2e/cases/lnp) package.
+The performance test cases are located in the [`teammates.lnp.cases`](https://github.com/TEAMMATES/teammates/tree/master/src/lnp/java/teammates/lnp/cases) package.
 
 ## Creating Performance Tests
