+++ src/lnp/java/teammates/lnp/cases/InstructorUpdateLNPTest.java
@@ -1,4 +1,4 @@
-package teammates.e2e.cases.lnp;
+package teammates.lnp.cases;
 
 import java.io.IOException;
 import java.time.Instant;
@@ -32,9 +32,9 @@ import teammates.common.datatransfer.questions.FeedbackTextQuestionDetails;
 import teammates.common.datatransfer.questions.FeedbackTextResponseDetails;
 import teammates.common.util.Const;
 import teammates.common.util.JsonUtils;
-import teammates.e2e.util.JMeterElements;
-import teammates.e2e.util.LNPSpecification;
-import teammates.e2e.util.LNPTestData;
+import teammates.lnp.util.JMeterElements;
+import teammates.lnp.util.LNPSpecification;
+import teammates.lnp.util.LNPTestData;
 import teammates.ui.request.InstructorCreateRequest;
 
 /**
