+++ src/lnp/java/teammates/lnp/cases/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Contains L&P test cases.
  */
-package teammates.e2e.cases.lnp;
+package teammates.lnp.cases;
