+++ src/lnp/java/teammates/lnp/util/LNPTestData.java
@@ -1,4 +1,4 @@
-package teammates.e2e.util;
+package teammates.lnp.util;
 
 import java.io.IOException;
 import java.util.HashMap;
