+++ src/lnp/java/teammates/lnp/util/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains infrastructure and helpers needed for running the L&P tests.
+ */
+package teammates.lnp.util;
