+++ static-analysis/teammates-pmdMain.xml
@@ -8,6 +8,7 @@
     </description>
     <exclude-pattern>.*/test/java/.*</exclude-pattern>
     <exclude-pattern>.*/e2e/java/.*</exclude-pattern>
+    <exclude-pattern>.*/lnp/java/.*</exclude-pattern>
     <exclude-pattern>.*/client/java/.*</exclude-pattern>
     <exclude-pattern>.*.html</exclude-pattern>
     <exclude-pattern>.*.xml</exclude-pattern>
